Move `is_conky_already_running()` to src/linux.cc.
